Improve detection so that the project is runnable locally and from another directory
