+++ .gitignore
@@ -1,4 +1,2 @@
 /venv/
 **/__pycache__/
-build/
-overcooked_sim.egg-info/
