+++ main.py
@@ -1,4 +1,4 @@
-import overcooked_utils
+import overcooked_env
 import argparse
 
 parser = argparse.ArgumentParser()
@@ -7,7 +7,7 @@ parser.add_argument("--seed", help="The seed to use for the environment.", defau
 args = parser.parse_args()
 
 noisy_randomization = True
-env, json = overcooked_utils.create_overcooked_env(args.environment_name, args.seed, noisy_randomization)
+env, json = overcooked_env.create_overcooked_env(args.environment_name, args.seed, noisy_randomization)
 obs, info = env.reset()
 env.render(mode='human')
 done = False
