+++ pddlgym_interface.py
@@ -2,7 +2,7 @@ import pddlgym
 import os
 import shutil
 import numpy as np
-from .overcooked_exceptions import OvercookedEnvironmentDoesNotExistException
+from overcooked_exceptions import OvercookedEnvironmentDoesNotExistException
 
 CURRENT_DIR_PATH = os.path.dirname(os.path.abspath(__file__))
 ENVIRONMENT_DIR_PATH = os.path.join(CURRENT_DIR_PATH, "environments")
