--- setup.cfg
@@ -1,6 +1,0 @@
-[metadata]
-name = overcooked_sim
-version = 0.1.0
-
-[options]
-packages = find:
